- Creada la primera version del paso a jvm (sin testear y a medias)    - Comentadas todas las instrucciones con lo que deben hacer. La instruccion suma puede servir de base para hacer el resto, pues es la unica de las complejas que esta implementada.    - He cambiado tambien los toChar, toNat, toInt y toFloat de los tipos de dato para adaptarlo a los castings que dice el enunciado de la practica.    - Haciendo esto anterior me di cuenta que no tenemos ninguna instruccion CastNat, que nos hace farta    - Al implementar las instrucciones intentad no usar el casting de los datos, aplicad las operaciones solo si los dos operando son del mismo tipo (como se hace en la suma). En el foro dice el tomatoma que es mejor hacerlo asi (cuanto menos inteligente deba ser el interprete mejor)

+++ trunk/src/pila/interprete/datos/DatoPila.java
@@ -1,6 +1,5 @@
 package pila.interprete.datos;
 
-import java.io.DataOutput;
 import java.io.DataOutputStream;
 import java.io.IOException;
 import pila.Dato;
@@ -45,7 +44,7 @@ public abstract class DatoPila extends Dato {
      * @throws DatoExc si son datos de tipos distintos y
      * el casting no puede hacerse
      */
-    public long toNatural() throws DatoExc {
+    public int toNatural() throws DatoExc {
         throw new DatoExc(this,toString()+" no puede transformarse en " +
                 "un natural");
     }
